A major change on the conversation flow management
Since there is no way to know where the conversation is when receiving
messages from the server, a 'process_message' function was added serving
as a gatekeeper to send the conversation to the stage it currently is
while is up and a user is interacting with it through whatsapp.

Also individual tests for each stage of the conversation are being added
at the bottom of the module.